+++ lib/file_hutch/rails.rb
@@ -5,7 +5,7 @@ require_relative "rails/attachable"
 require_relative "backfill"
 
 module FileHutch
-  def self.deprecator = @deprecator ||= ActiveSupport::Deprecation.new("0.3", "file_hutch")
+  def self.deprecator = @deprecator ||= ActiveSupport::Deprecation.new("0.4", "file_hutch")
 end
 
 require_relative "rails/engine"